simple_upload/sam-dev/mysam/upload_function/requirements.txt
pyjwt
urllib3==1.26